requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pytest==7.4.3
python-socketio==5.10.0
aiohttp==3.9.1
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.25
langchain-core==0.2.40
langsmith==0.1.129
chromadb==0.4.18
beautifulsoup4==4.12.2
openai==1.51.0
numpy==1.24.3
scipy==1.11.4
redis==5.0.1
httpx==0.24.1
faiss-cpu==1.7.4